+++ Package/lib/src/Widgets/Views/Landing/LandingPageView.dart
@@ -219,15 +219,18 @@ class _LandingPageViewState extends State<LandingPageView> {
       ],
     );
 
-    var halfGlassPane = FloatingContainerElement(
-      child: Container(
-        height: screenSize.height,
-        width: screenSize.width * 0.50,
-        decoration: ButtonBackingDecoration(
-          variant: buttonDecorationVariants.edgedRectangle,
-          decorationVariant: decorationPriority.inactive,
-        ).buildBacking(),
-      ),
+    var halfGlassPane = Container(
+      height: screenSize.height,
+      width: screenSize.width * 0.50,
+      decoration: BoxDecoration(
+          gradient: LinearGradient(
+        begin: Alignment.centerLeft,
+        end: Alignment.centerRight,
+        colors: [
+          coloration.sameColor(),
+          coloration.sameColor().withOpacity(0.0),
+        ],
+      )),
     );
 
     return Scaffold(