+++ Package/test/Object Tests/Element Obj Tests/aureus_object_elements_test.dart
@@ -1,3 +1,0 @@
-part of aureus_tests;
-
-class AureusObjectElementsTests {}